+++ src/lib.rs
@@ -11,6 +11,7 @@
 
 pub mod api;
 pub mod conn;
+pub mod emoji;
 mod huehash;
 mod replies;